Fix desciption and author meta tags
The tags were using the wrong syntax to insert the variables. I also
wrapped them in conditionals so that they do not add the tags if the
variables are empty (not defined).

Link to Google’s font API changed to https.

I thought of mirroring that Google font locally, but it links to a
stylesheet that links to several versions of the font. Too much manual
work. The site will still look fine if Google eventually takes that API
offline, so I will not bother.